Move type to shared as it is used by pdf and csv flow
